requirements.txt
flask
flask-sqlalchemy
flask-cors
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-groq==0.2.1
langchain-openai==0.3.0
langchain-anthropic==0.3.1
langchain-cohere==0.3.1
langchain-huggingface==0.1.2
faiss-cpu
sentence-transformers
pypdf
python-docx
pandas
openpyxl
pytesseract
Pillow
python-dotenv
redis
cohere
tiktoken
sqlalchemy
werkzeug
requests
numpy
